Added three new functions

+++ admin/myblogposts.php
@@ -8,31 +8,58 @@ include_once "../constants.php";
 
 $username = $_SESSION['username'];
 $userid = $_SESSION['id'];
+$message = "";
+
 include_once "../controller/db_controller.php";
 include_once "../model/message.php";
 
+
 if(isset($_GET['edit'])){
+
 $postid = $_GET['edit'];
+$author = getUsernameFromPostID($postid);
+//check if post belongs to user
+    if($username != $author){
+        $message = unauthorizedAuthorMsg();
+        header("Location: myblogposts.php");
+        die("redirecting to myblogposts.php");
+
+    }else{
+
+
 $post_title = getPostTitleByID($postid);
 $post_content = getPostContentByID($postid);
 $post_image = getImageByPostId($postid);
-console_log($post_content);
+
+
+
+    }
 }
 
 //check if get request = delete
-if(isset($_GET['delete'])){
-    console_log("inne i GET DELETE");
+if(isset($_GET['delete'])) {
+   //post id
     $postid = $_GET['delete'];
-    console_log("före deletepost");
-    deletePostFromDatabase($postid);
-    $message = deletePostMsg();
-    console_log("Message i delete: $message");
-    console_log("efter deletepost");
-    //header("Location:myblogposts.php");
+    //get username from postid
+    $author = getUsernameFromPostID($postid);
+    //check if logged in user is the same as author for the postid.
+    if ($username != $author) {
 
+        header("Location: myblogposts.php");
+        die("redirecting to myblogposts.php");
+        $message = unauthorizedAuthorMsg();
+
+    } else {
+        //delete the post from database
+        deletePostFromDatabase($postid);
+        $message = deletePostMsg();
+    }
 }
 
 if(isset($_POST['edit-load-blog-post'])) {
+    
+
+
     $post_title = $_POST['edit-title'];
     $post_content = $_POST['edit-text-area'];
     $post_image = "";
@@ -44,10 +71,10 @@ if(isset($_POST['edit-load-blog-post'])) {
         $imageName = $_FILES['edit-image']['name'];
         $upload_status = 1;
         $image_status = 1;
-        console_log("inne i is_uploaded_file");
+
         //check if the uploaded file exists on the server
         if (file_exists("../uploads/" . $imageName)) {
-            console_log("inne i file_exists");
+
             //if it exists send error message.
             $message = renameImageFileMsg();
             $upload_status = 0;
@@ -61,26 +88,29 @@ if(isset($_POST['edit-load-blog-post'])) {
         $image_status = 0;
     }
 
-    if ($image_status == 1) {
-        console_log("imagestatus = 1");
-        if(doesImageExistInPost($postid) == 1){
-            console_log("imageexist in post = 1");
-            move_uploaded_file($tmp_file, "../uploads/" . $imageName);
-            updateImage($postid,$imageName);
-        }else{
-            console_log("imageexist != 1");
-            console_log("Du har ingen fil uppladdad i databasen");
-            insertImageToDatabase($postid,$imageName);
+    //if upload status is ok update post.
+    if ($upload_status == 1) {
+
+        updatePost($postid, $post_title, $post_content, $date);
+
+        $message = editPostSuccessMsg();
+
+        //if an image is uploaded
+        if ($image_status == 1) {
+            //checks if image exist in post,
+            if(doesImageExistInPost($postid) == 1){
+                //move image to directory
+                move_uploaded_file($tmp_file, "../uploads/" . $imageName);
+                updateImage($postid,$imageName);
+            }else{
+
+                //insert image name into db
+                move_uploaded_file($tmp_file, "../uploads/" . $imageName);
+                insertImageToDatabase($postid,$imageName);
+
+            }
 
         }
-
-    }
-    if ($upload_status == 1) {
-        console_log("upload status = 1");
-        updatePost($postid, $post_title, $post_content, $date);
-        echo("uppladdad!!");
-        $message = editPostSuccessMsg();
-        console_log("message efter upload: ".$message);
     }
 
 }
@@ -114,8 +144,14 @@ if(isset($_POST['edit-load-blog-post'])) {
     <div class="col-8">
         <?php
 
-        echo($message);
-        console_log("message i början av sidan: " .$message);
+        if(isset($_POST['edit-load-blog-post'])) {
+            echo($message);
+        }
+
+        if(isset($_GET['delete'])) {
+            echo($message);
+        }
+
         ?>
         <h2> Mina inlägg </h2>
         <form method="post">
@@ -127,7 +163,7 @@ if(isset($_POST['edit-load-blog-post'])) {
                <tr>
                    <th>Titel</th>
 
-                   <th colspan="2">Action</th>
+                   <th colspan="2"></th>
                </tr>
 
                </thead>
